# backend/src/main/java/com/game/backend/admin/api/AdminItemOperationRequest.java
package com.game.backend.admin.api;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;

import java.util.Map;
import java.util.UUID;

/**
 * Машиночитаемая команда для явных admin item endpoints из ТЗ.
 */
public record AdminItemOperationRequest(
    @NotNull
    UUID playerId,

    @NotBlank
    String itemId,

    @NotNull
    Long catalogVersion,

    @NotBlank
    String reason,

    String disabledReason,

    String unlockHintCode,

    Map<String, Object> unlockHintPayload
) {
}